Bobs equipment (#989)
* Add check before replacing tech

By default, electric engine is blue science. However a couple of Bob's mods move it to green science. In this case, angels-components-mechanical-4 is too high a tier prerequisite.

* Adjust equipment/armor tech tiers

* Adjust equipment/armor tech tiers

* Adjust equipment/armor tech tiers

Low density structures are are a higher tech tier

* Reverting previous change

* Burner Phase setting can be enabled without Burner Assemblers

* Unit test: mod settings change

Removed duplicate setting
Moved setting to correct heading

* Unit test update: Bob modules

Modules can be considered same tier as when they are unlocked. No need for them to be after

* Armor ingredients and prerequisites