Fix artifact name selection
Depending on present mods, either the artifacts added by Bob's Enemies
or the ones added by Angel's Exploration are used.

This fixes the check also checking for bobsplates. That mod is only relevant
for adding the tier 4 gem to the recipe, not for picking what item names to
use for artifacts.

Note: This can probably be cleaned up by having a saner mod
cross-integration approach with mods agreeing upon what gets generated
based on the settings stage, instead of checking the results of the data
stage and running all of this logic in the data-update stage.