angelsrefining: Remove enemy multiplier slider
This slider is a strange way to enable/disable enemies.
Instead, offer a preset without enemies and point the player towards it
with the welcome message.

The slider-based implementation also broke achievement lockout - where
some achievements are locked on playthroughs with less or no enemies.

For 2.0 the prior implementation of overriding spawner map generation to
never generate spawners will also conflict with space age's progression,
where capturing a spawner is part of the tech tree.